% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convoSPAT_paramEst.R
\name{anis_model_kappa}
\alias{anis_model_kappa}
\title{Local parameter estimation for the nonstationary model.}
\usage{
anis_model_kappa(params)
}
\arguments{
\item{params}{A vector of inputs: lam1, lam2, eta, tausq, sigmasq, and nu}
}
\value{
This function returns the loglikelihood for a particular
combination of (lam1, lam2, eta, tausq, sigmasq, nu).
}
\description{
This function calculates the maximum likelihood estimates of
local variance parameters lam1, lam2, eta, tausq, sigmasq, and
nu (smoothness) using a Gaussian likelihood with
an anisotropic covariance structure.
}
\examples{
\dontrun{
anis_model_kappa( c(1,1,0,0.1,1,1) )
}
}

