% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convoSPAT_fitpred.R
\name{Aniso_fit}
\alias{Aniso_fit}
\title{Fit the stationary spatial model}
\usage{
Aniso_fit(geodata, coords = geodata$coords, data = geodata$data,
  cov.model = "exponential", mean.model = data ~ 1, iso.model = FALSE)
}
\arguments{
\item{geodata}{A list containing elements \code{coords} and \code{data} as
described next. Typically an object of the class "\code{geodata}", although
a geodata object only allows \code{data} to be a vector (no replicates).
If not provided, the arguments \code{coords} and \code{data} must be
provided instead.}

\item{coords}{An N x 2 matrix where each row has the two-dimensional
coordinates of the N data locations. By default, it takes the \code{coords}
component of the argument \code{geodata}, if provided.}

\item{data}{A vector or matrix with N rows, containing the data values.
By default, it takes the \code{data} component of the argument
\code{geodata}, if provided. The stationary model cannot take multiple
replicates.}

\item{cov.model}{A string specifying the model for the correlation
function; following \code{geoR}, defaults to \code{"exponential"}.
Options available in this package are: "\code{exponential}",
\code{"cauchy"}, \code{"matern"}, \code{"circular"}, \code{"cubic"},
\code{"gaussian"}, \code{"spherical"}, and \code{"wave"}. For further
details, see documentation for \code{\link[geoR]{cov.spatial}}.}

\item{mean.model}{An object of class \code{\link[stats]{formula}},
specifying the mean model to be used. Defaults to an intercept only.}

\item{iso.model}{Logical; indicates whether the stationary model should
be isotropic (\code{TRUE}) or anisotropic (\code{FALSE}). Defaults to
\code{FALSE}.}
}
\value{
A list with the following components:
\item{MLEs.save}{Table of global maximum likelihood estimates for lam1,
lam2, eta, tausq, sigmasq, and nu.}
\item{data}{Observed data values.}
\item{beta.GLS}{Vector of generalized least squares estimates of beta,
the mean coefficients.}
\item{aniso.pars}{Vector of MLEs for the anisotropy parameters lam1,
lam2, eta.}
\item{aniso.mat}{2 x 2 anisotropy matrix, calculated from
\code{aniso.pars}.}
\item{tausq.est}{Scalar maximum likelihood stimate of tausq (nugget
variance).}
\item{sigmasq.est}{Scalar maximum likelihood estimate of sigmasq (process
variance).}
\item{kappa.MLE}{Scalar maximum likelihood estimate for kappa (when
applicable).}
\item{cov.model}{String; the correlation model used for estimation.}
\item{coords}{N x 2 matrix of observation locations.}
\item{global.loglik}{Scalar value of the maximized likelihood from the
global optimization (if available).}
\item{likfit.obj}{Object output from the \code{\link[geoR]{likfit}}
function; used in calculating predictions.}
}
\description{
\code{Aniso_fit} estimates the parameters of the stationary spatial model.
Required inputs are the observed data and locations (a geoR object
with $coords and $data). Optional inputs include the covariance model
(exponential is the default), and whether or not the model should be
isotropic. The heavy lifting of this function is done by the
\code{\link[geoR]{likfit}} function in \code{geoR}.
}
\examples{
\dontrun{
# Using the "simdata" data available in the package.
anisofit.model <- Aniso_fit( coords = simdata$sim.locations,
data = simdata$sim.data, cov.model = "exponential",
mean.model = simdata$sim.data ~ simdata$sim.locations
+ simdata$sim.locations[-holdout.index,2] )
}
}

