% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convoSPAT_fitpred.R
\name{Aniso_pred}
\alias{Aniso_pred}
\title{Obtain predictions at unobserved locations for the stationary
spatial model.}
\usage{
Aniso_pred(Aniso.fit.obj, pred.coords, mean.model.d = ~1, mean.model.l = ~1)
}
\arguments{
\item{Aniso.fit.obj}{Object from calling \code{Aniso.fit}.}

\item{pred.coords}{Matrix of locations where predictions are required.}

\item{mean.model.d}{Formula object; following
\code{\link[geoR]{krige.conv}}, specifies
the trend (covariate) values at the data locations.}

\item{mean.model.l}{Formula object; following
\code{\link[geoR]{krige.conv}}, specifies
the trend (covariate) values at the prediction locations.}
}
\value{
A list with the following components:
\item{pred.means}{Vector of the kriging predictor, for each location in
\code{pred.coords}.}
\item{pred.SDs}{Vector of the kriging standard errors, for each location
in \code{pred.coords}.}
}
\description{
\code{Aniso_pred} calculates the kriging predictor and corresponding
standard errors at unmonitored sites.
}
\examples{
\dontrun{
pred.S <- Aniso_pred( Aniso.fit.obj = anisofit.model,
pred.coords = matrix(c(1,1), ncol=2),
mean.model.d = simdata$sim.data ~ simdata$sim.locations[,1]
+ simdata$sim.locations[,2],
mean.model.l = ~ c(1) + c(1) )
}
}

