% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convoSPAT_simulate.R
\name{kernel_cov}
\alias{kernel_cov}
\title{Calculate a kernel covariance matrix.}
\usage{
kernel_cov(params)
}
\arguments{
\item{params}{A vector of three parameters, corresponding to
(lam1, lam2, eta). The eigenvalues (lam1 and lam2) must be positive.}
}
\value{
A 2 x 2 kernel covariance matrix.
}
\description{
\code{kernel_cov} calculates a 2 x 2 matrix based on the eigendecomposition
components (two eigenvalues and angle of rotation).
}
\examples{
kernel_cov(c(1, 2, pi/3))
}

