% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convoSPAT_paramEst.R
\name{make_global_loglik1}
\alias{make_global_loglik1}
\title{Constructor functions for global parameter estimation.}
\usage{
make_global_loglik1(data, Xmat, Corr, nugg2.var)
}
\arguments{
\item{data}{A vector or matrix of data to use in the likelihood
calculation.}

\item{Xmat}{The design matrix for the mean model.}

\item{Corr}{The correlation matrix.}

\item{nugg2.var}{Fixed values for the variance of the second nugget term.}
}
\value{
This function returns another function for use in \code{optim}.
}
\description{
This function generates another function to be used within \code{optim} to
obtain maximum likelihood estimates of
global variance parameters tausq, sigmasq with a fixed correlation
matrix (smoothness is fixed).
}
\examples{
\dontrun{
make_global_loglik1( data, Xmat, Dmat, Vmat )
}

}

