% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convoSPAT_paramEst.R
\name{make_global_loglik3_kappa}
\alias{make_global_loglik3_kappa}
\title{Constructor functions for global parameter estimation.}
\usage{
make_global_loglik3_kappa(data, Xmat, cov.model, Scalemat, Distmat,
  obs.variance, nugg2.var)
}
\arguments{
\item{data}{A vector or matrix of data to use in the likelihood
calculation.}

\item{Xmat}{The design matrix for the mean model.}

\item{cov.model}{String; the covariance model.}

\item{Scalemat}{Matrix; contains the scaling quantities from the
covariance function.}

\item{Distmat}{Matrix; contains the scaled distances.}

\item{obs.variance}{A vector containing the spatially-varying variance
corresponding to each data location.}

\item{nugg2.var}{Fixed values for the variance of the second nugget term.}
}
\value{
This function returns another function for use in \code{optim}.
}
\description{
This function generates another function to be used within \code{optim} to
obtain maximum likelihood estimates of
global variance parameters tausq and nu. The process variance is
taken to be spatially-varying.
}
\examples{
\dontrun{
make_global_loglik3_kappa( data, Xmat, cov.model, Scalemat, Distmat, obs.variance )
}

}

