% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convoSPAT_paramEst.R
\name{overall_lik1}
\alias{overall_lik1}
\title{Global parameter estimation for the nonstationary model.}
\usage{
overall_lik1(params)
}
\arguments{
\item{params}{A vector of inputs: tausq and sigmasq.}
}
\value{
This function returns the loglikelihood for a particular
combination of (tausq, sigmasq).
}
\description{
This function calculates the maximum likelihood estimates of
global variance parameters tausq, sigmasq with a fixed correlation
matrix (smoothness is fixed).
}
\examples{
\dontrun{
overall_lik1( c(0.5,1) )
}
}

