% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convoSPAT_paramEst.R
\name{overall_lik1_kappa}
\alias{overall_lik1_kappa}
\title{Global parameter estimation for the nonstationary model.}
\usage{
overall_lik1_kappa(params)
}
\arguments{
\item{params}{A vector of inputs: tausq, sigmasq, and nu.}
}
\value{
This function returns the loglikelihood for a particular
combination of (tausq, sigmasq, nu).
}
\description{
This function calculates the maximum likelihood estimates of
global variance parameters tausq, sigmasq, and nu.
}
\examples{
\dontrun{
overall_lik1_kappa( c(0.5,1,1) )
}
}

