% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convoSPAT_paramEst.R
\name{overall_lik2}
\alias{overall_lik2}
\title{Global parameter estimation for the nonstationary model.}
\usage{
overall_lik2(params)
}
\arguments{
\item{params}{Scalar input for sigmasq.}
}
\value{
This function returns the loglikelihood for a particular
value of sigmasq.
}
\description{
This function calculates the maximum likelihood estimates of
global variance parameter sigmasq with a fixed correlation
matrix (smoothness is fixed). The nugget variance is taken
to be spatially-varing.
}
\examples{
\dontrun{
overall_lik2( 1 )
}
}

