% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convoSPAT_paramEst.R
\name{overall_lik2_kappa}
\alias{overall_lik2_kappa}
\title{Global parameter estimation for the nonstationary model.}
\usage{
overall_lik2_kappa(params)
}
\arguments{
\item{params}{A vector of inputs: sigmasq and nu.}
}
\value{
This function returns the loglikelihood for a particular
combination of (sigmasq, nu).
}
\description{
This function calculates the maximum likelihood estimates of
global variance parameters sigmasq and nu. The nugget variance is
taken to be spatially-varying.
}
\examples{
\dontrun{
overall_lik2_kappa( c(1,1) )
}
}

