% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convoSPAT_paramEst.R
\name{overall_lik3_kappa}
\alias{overall_lik3_kappa}
\title{Global parameter estimation for the nonstationary model.}
\usage{
overall_lik3_kappa(params)
}
\arguments{
\item{params}{A vector of inputs: tausq and nu.}
}
\value{
This function returns the loglikelihood for a particular
combination of (tausq, nu).
}
\description{
This function calculates the maximum likelihood estimates of
global variance parameters tausq and nu. The process variance is
taken to be spatially-varying.
}
\examples{
\dontrun{
overall_lik3_kappa( c(0.5,1) )
}
}

