% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convoSPAT_paramEst.R
\name{overall_lik4_kappa}
\alias{overall_lik4_kappa}
\title{Global parameter estimation for the nonstationary model.}
\usage{
overall_lik4_kappa(params)
}
\arguments{
\item{params}{A vector of inputs: nu.}
}
\value{
This function returns the loglikelihood for a particular
combination of (nu).
}
\description{
This function calculates the maximum likelihood estimates of
global variance parameters nu. The process variance
and nugget variance are taken to be spatially-varying.
}
\examples{
\dontrun{
overall_lik4_kappa( c(1) )
}
}

