% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convoSPAT_summplot.R
\name{plot.Aniso}
\alias{plot.Aniso}
\title{Plot of the estimated correlations from the stationary model.}
\usage{
\method{plot}{Aniso}(x, ref.loc = NULL, all.pred.locs = NULL, grid = TRUE,
  ...)
}
\arguments{
\item{x}{An "Aniso" object, from Aniso_fit().}

\item{ref.loc}{Vector of length 2; the reference location.}

\item{all.pred.locs}{A matrix of all prediction locations.}

\item{grid}{Logical; indicates if the \code{all.pred.locs}
are on a rectangular grid (\code{TRUE}) or not (\code{FALSE}).}

\item{...}{Arguments passed to \code{plot} functions.}
}
\value{
A plot of either the estimated ellipses or estimated
correlation is printed.
}
\description{
This function plots the estimated correlation between a reference
point and all other prediction locations.
}
\examples{
\dontrun{
plot.Aniso( Aniso.object )
}

}

