% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convoSPAT_summplot.R
\name{plot.NSconvo}
\alias{plot.NSconvo}
\title{Plot from the nonstationary model: either (1) estimated anisotropy
ellipses, or (2) estimated correlations.}
\usage{
\method{plot}{NSconvo}(x, plot.ellipses = TRUE, fit.radius = NULL,
  aniso.mat = NULL, true.mc = NULL, ref.loc = NULL,
  all.pred.locs = NULL, grid = TRUE, ...)
}
\arguments{
\item{x}{A "NSconvo" object, from NSconvo_fit().}

\item{plot.ellipses}{Logical; indicates whether the estimated
ellipses should be plotted (\code{TRUE}) or estiamted correlations
(\code{FALSE}).}

\item{fit.radius}{Scalar; defines the fit radius used for the local
likelihood estimation.}

\item{aniso.mat}{2 x 2 matrix; contains the estimated anisotropy
ellipse from the stationary model (for comparison).}

\item{true.mc}{The true mixture component ellipses, if known.}

\item{ref.loc}{Vector of length 2; the reference location.}

\item{all.pred.locs}{A matrix of all prediction locations.}

\item{grid}{Logical; indicates if the \code{all.pred.locs}
are on a rectangular grid (\code{TRUE}) or not (\code{FALSE}).}

\item{...}{Other options passed to \code{plot}.}
}
\value{
A plot of either the estimated ellipses or estimated
correlation is printed.
}
\description{
This function plots either the estimated anisotropy ellipses for each
of the mixture component locations of the estimated correlation
between a reference point and all other prediction locations.
}
\examples{
\dontrun{
plot.NSconvo( NSconvo.object )
}
}

