% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convoSPAT_summplot.R
\name{plot_ellipses}
\alias{plot_ellipses}
\title{Plot estimated anisotropy ellipses.}
\usage{
plot_ellipses(locations, kernels, length)
}
\arguments{
\item{locations}{A N x 2 matrix of observed locations.}

\item{kernels}{A N x 2 x 2 array of estimated anisotropy ellipses
corresponding to the observed locations.}

\item{length}{Scalar; how many ellipses should be plotted in
each direction.}
}
\value{
A plot of estimated ellipses is printed.
}
\description{
This function plots a subset of the estimated anisotropy ellipses
over the spatial domain. This function works only when the locations
are on a grid.
}
\examples{
\dontrun{
plot_ellipses( locations = simdata$coords,
kernels = fit.model$kernel.ellipses, length = 4 )
}
}

