% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convoSPAT_summplot.R
\name{plot_mc}
\alias{plot_mc}
\title{Plot estimated anisotropy ellipses.}
\usage{
plot_mc(mc.locations, mc.kernels, fit.radius, aniso.mat = NULL)
}
\arguments{
\item{mc.locations}{A matrix of mixture component locations.}

\item{mc.kernels}{An array of kernel matrices corresponding to each
of the mixture component locations.}

\item{fit.radius}{Scalar; defines the fit radius used for the local
likelihood estimation.}

\item{aniso.mat}{2 x 2 matrix; contains the estimated anisotropy
ellipse from the stationary model (for comparison).}
}
\value{
A plot of the estimated ellipses is printed.
}
\description{
This function plots the estimated anisotropy ellipses for each
of the mixture component locations.
}
\examples{
\dontrun{
plot.mc( mc.locations = simdata$mc.locations,
mc.kernels = fit.model$mc.kernels, fit.radius = 1 )
}
}

