% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convoSPAT_summplot.R
\name{summary.Aniso}
\alias{summary.Aniso}
\title{Summarize the stationary model fit.}
\usage{
\method{summary}{Aniso}(object, ...)
}
\arguments{
\item{object}{An "Aniso" object, from \code{Aniso_fit}.}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
Text containing the model fitting results.
}
\description{
\code{summary.Aniso} prints relevant output from the model fitting
procedure.
}
\examples{
\dontrun{
summary.Aniso( Aniso.object )
}

}

