% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convoSPAT_summplot.R
\name{summary.NSconvo}
\alias{summary.NSconvo}
\title{Summarize the nonstationary model fit.}
\usage{
\method{summary}{NSconvo}(object, ...)
}
\arguments{
\item{object}{A "NSconvo" object, from \code{NSconvo_fit}.}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
Text containing the model fitting results.
}
\description{
\code{summary.NSconvo} prints relevant output from the model fitting
procedure.
}
\examples{
\dontrun{
summary.NSconvo( NSconvo.object )
}
}

