% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convoSPAT_summplot.R
\name{summary_NS}
\alias{summary_NS}
\title{Summarize the nonstationary model fit.}
\usage{
summary_NS(fit.model)
}
\arguments{
\item{fit.model}{An object which is the output of
\code{NSconvo.fit}.}
}
\value{
Text containing the model fitting results.
}
\description{
\code{summary_NS} prints relevant output from the model fitting
procedure.
}
\examples{
\dontrun{
summary_NS( fit.model )
}
}

