% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{US.mc.grids}
\alias{US.mc.grids}
\title{Mixture component grids for the western United States}
\format{A list with two elements:
\describe{
   \item{Element 1}{Coarse mixture component grid.}
   \item{Element 2}{Fine mixture component grid.}

}}
\usage{
US.mc.grids
}
\description{
A list of two mixture component grids for fitting the nonstationary
model to the western United States precipitation data.
}
\keyword{datasets}
