% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{US.prediction.locs}
\alias{US.prediction.locs}
\title{Prediction locations for the western United States}
\format{
A matrix with two columns:
\describe{
   \item{Column 1}{Longitude of the prediction grid.}
   \item{Column 2}{Latitude of the prediction grid.}

}
}
\usage{
US.prediction.locs
}
\description{
A matrix with two columns containing a fine grid of locations for
which to make a filled-in prediction map for the western United States.
}
\keyword{datasets}
