% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{USprecip}
\alias{USprecip}
\title{Annual precipitation measurements from the western United States, 1997}
\format{A data frame with the following variables:
\describe{
   \item{longitude}{Longitude of the monitoring site.}
   \item{latitude}{Latitude of the monitoring site.}
   \item{annual.ppt}{Annual precipitation for the monitoring site, in
   millimeters.}
   \item{log.annual.ppt}{Annual precipitation for the monitoring site, in
   log millimeters.}
   \item{ind.holdout}{Indicates a hold-out sample of the monitoring sites.}

}}
\source{
\url{http://www.image.ucar.edu/GSP/Data/US.monthly.met/}
}
\usage{
USprecip
}
\description{
A data set containing the annual precipitation for almost 1500 locations
in the western United States.
}
\keyword{datasets}

